Rf: Add redux(store,reducer,actions), restructure navigation, etc ...
