--- src/components/MapContainer/MapContainerStyles.js
@@ -1,12 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = {
-   container: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-   },
-   map: {
-      ...StyleSheet.absoluteFillObject,
-   },
-};
