+++ src/navigation/NavigationStack.js
@@ -1,46 +1,47 @@
-// import * as React from 'react';
-// import { NavigationContainer } from '@react-navigation/native';
-// import { createStackNavigator } from '@react-navigation/stack';
-// import { useSelector } from 'react-redux';
-
-// import { navigationRef } from './NavigationService';
-
-// import Login from 'app/screens/Login';
-// import Home from 'app/screens/Home';
-
-// const Stack = createStackNavigator();
-
-// const homeOptions = {
-//   title: 'My home',
-//   headerStyle: {
-//     backgroundColor: '#f4511e',
-//   },
-//   headerTintColor: '#fff',
-//   headerTitleStyle: {
-//     fontWeight: 'bold',
-//   },
-// };
-
-// function App() {
-//   const isLoggedIn = useSelector(state => state.loginReducer.isLoggedIn);
-
-//   return (
-//     <NavigationContainer ref={navigationRef}>
-//       <Stack.Navigator>
-//         {isLoggedIn ? (
-//           <Stack.Screen name="Home" component={Home} options={homeOptions} />
-//         ) : (
-//           <Stack.Screen
-//             name="Login"
-//             component={Login}
-//             options={{
-//               animationTypeForReplace: isLoggedIn ? 'push' : 'pop',
-//             }}
-//           />
-//         )}
-//       </Stack.Navigator>
-//     </NavigationContainer>
-//   );
-// }
-
-// export default App;
+import { createAppContainer } from 'react-navigation';
+import { createStackNavigator } from 'react-navigation-stack';
+import RiderHome from '../screens/rider/RiderHome';
+import RiderLogin from '../screens/rider/RiderLogin';
+import RiderRegister from '../screens/rider/RiderRegister';
+import RiderVerifyNumber from '../screens/rider/RiderVerifyNumber';
+import RiderForgotPassword from '../screens/rider/RiderForgotPassword';
+import DriverHome from '../screens/driver/DriverHome';
+import DriverLogin from '../screens/driver/DriverLogin';
+import DriverRegister from '../screens/driver/DriverRegister';
+import AuthLoadingScreen from '../screens/main/AuthLoadingScreen';
+import RiderDriverScreenChoice from '../screens/main/RiderDriverScreenChoice';
+import DriverRegLog from '../screens/driver/DriverRegLog';
+import RiderRegLog from '../screens/rider/RiderRegLog';
+import RiderHomeContents from '../screens/rider/RiderHomeContents';
+import RiderPickUp from '../screens/rider/RiderPickUp';
+const AuthStackRider = createStackNavigator({
+   Rider: { screen: RiderHome },
+});
+const AuthStackDriver = createStackNavigator({
+   Driver: { screen: DriverHome },
+});
+const AuthStack = createAppContainer(
+   createStackNavigator({
+      Home: { screen: RiderHome },
+      RiderScreen: { screen: RiderRegLog },
+      RiderReg: { screen: RiderRegister },
+      RiderVerifyNum: { screen: RiderVerifyNumber },
+      RiderResetPassWord: { screen: RiderForgotPassword },
+      DriverScreen: { screen: DriverRegLog },
+      DriverLog: { screen: DriverLogin },
+      DriverReg: { screen: DriverRegister },
+   }),
+);
+const RiderHomeStackNav = createStackNavigator(
+   {
+      Main: { screen: RiderHomeContents },
+      pickUpLocation: { screen: RiderPickUp },
+   },
+   {
+      headerMode: 'none',
+      navigationOptions: {
+         headerVisible: false,
+      },
+   },
+);
+export { AuthStack, AuthStackDriver, AuthStackRider , RiderHomeStackNav };
