+++ src/navigation/index.js
@@ -1,2 +1,4 @@
-// import NavigationStack from './NavigationStack';
-// export default NavigationStack;
+import NavigationSwitch from './NavigationSwitch';
+//import RiderHomeDrawer from './NavigationDrawer';
+
+export default NavigationSwitch ;
