+++ src/redux/actions/navigationActions.js
@@ -1,7 +1,7 @@
 /*
  * Reducer actions related with navigation
  */
-import NavigationService from 'app/navigation/NavigationService';
+import NavigationService from '../navigation/NavigationService';
 
 export function navigateToHome(params) {
   NavigationService.navigate('Home', params);
