+++ src/redux/sagas/loginSaga.js
@@ -7,7 +7,7 @@
 import { put, call, select } from 'redux-saga/effects';
 import { delay } from 'redux-saga';
 import { Alert } from 'react-native';
-import * as loginActions from 'src/redux/actions/loginActions';
+import * as loginActions from '../actions/loginActions';
 export default function* loginAsync() {
   yield put(loginActions.enableLoader());
   const response = { success: true, data: { id: 1 } };
